# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MineOps Planning Service API
  version: 1.0.0
  description: API para gestionar actividades operacionales mineras.
servers:
  - url: http://localhost:8000/api/v1
tags:
  - name: Health
  - name: Activities
paths:
  /health:
    get:
      tags: [Health]
      summary: Consultar el estado del servicio
      operationId: healthCheck
      responses:
        '200':
          description: Servicio disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /activities:
    post:
      tags: [Activities]
      summary: Crear una actividad
      operationId: createActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityRequest'
      responses:
        '201':
          description: Actividad creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [Activities]
      summary: Listar actividades
      operationId: listActivities
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Priority'
        - $ref: '#/components/parameters/ResponsibleIdFilter'
        - $ref: '#/components/parameters/ScheduledFrom'
        - $ref: '#/components/parameters/ScheduledTo'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Direction'
      responses:
        '200':
          description: Página de actividades
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedActivitiesResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /activities/pending:
    get:
      tags: [Activities]
      summary: Listar actividades pendientes
      operationId: listPendingActivities
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Priority'
        - $ref: '#/components/parameters/ScheduledFrom'
        - $ref: '#/components/parameters/ScheduledTo'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Direction'
      responses:
        '200':
          description: Página de actividades pendientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedActivitiesResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /activities/overdue:
    get:
      tags: [Activities]
      summary: Listar actividades vencidas no finalizadas
      operationId: listOverdueActivities
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Priority'
        - $ref: '#/components/parameters/ResponsibleIdFilter'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Direction'
      responses:
        '200':
          description: Página de actividades vencidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedActivitiesResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /activities/responsible/{responsibleId}:
    get:
      tags: [Activities]
      summary: Listar actividades de un responsable
      operationId: listActivitiesByResponsible
      parameters:
        - name: responsibleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Priority'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Direction'
      responses:
        '200':
          description: Página de actividades del responsable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedActivitiesResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /activities/{id}:
    parameters:
      - $ref: '#/components/parameters/ActivityId'
    get:
      tags: [Activities]
      summary: Consultar una actividad
      operationId: getActivity
      responses:
        '200':
          description: Actividad encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Activities]
      summary: Actualizar una actividad
      operationId: updateActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActivityRequest'
      responses:
        '200':
          description: Actividad actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /activities/{id}/start:
    patch:
      tags: [Activities]
      summary: Iniciar una actividad
      operationId: startActivity
      parameters:
        - $ref: '#/components/parameters/ActivityId'
      responses:
        '200':
          description: Actividad iniciada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /activities/{id}/complete:
    patch:
      tags: [Activities]
      summary: Completar una actividad
      operationId: completeActivity
      parameters:
        - $ref: '#/components/parameters/ActivityId'
      responses:
        '200':
          description: Actividad completada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /activities/{id}/cancel:
    patch:
      tags: [Activities]
      summary: Cancelar una actividad
      operationId: cancelActivity
      parameters:
        - $ref: '#/components/parameters/ActivityId'
      responses:
        '200':
          description: Actividad cancelada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    ActivityId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 15
    Status:
      name: status
      in: query
      schema:
        $ref: '#/components/schemas/ActivityStatus'
    Priority:
      name: priority
      in: query
      schema:
        $ref: '#/components/schemas/ActivityPriority'
    ResponsibleIdFilter:
      name: responsible_id
      in: query
      schema:
        type: string
        format: uuid
    ScheduledFrom:
      name: scheduled_from
      in: query
      schema:
        type: string
        format: date
    ScheduledTo:
      name: scheduled_to
      in: query
      schema:
        type: string
        format: date
    Sort:
      name: sort
      in: query
      schema:
        type: string
        enum: [created_at, scheduled_date, due_date, title, priority, status]
        default: created_at
    Direction:
      name: direction
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
  responses:
    NotFound:
      description: Actividad o ruta no encontrada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflicto de estado o de datos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Datos inválidos o regla de negocio incumplida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Error inesperado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ActivityStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, COMPLETED, CANCELLED]
    ActivityPriority:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    Activity:
      type: object
      required:
        - id
        - code
        - title
        - description
        - area
        - location
        - responsible_id
        - priority
        - status
        - scheduled_date
        - due_date
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 50
        title:
          type: string
          maxLength: 200
        description:
          type: [string, 'null']
        area:
          type: string
        location:
          type: string
        responsible_id:
          type: string
          format: uuid
        priority:
          $ref: '#/components/schemas/ActivityPriority'
        status:
          $ref: '#/components/schemas/ActivityStatus'
        scheduled_date:
          type: string
          format: date
        due_date:
          type: [string, 'null']
          format: date
    CreateActivityRequest:
      type: object
      required: [code, title, area, location, responsible_id, priority, scheduled_date]
      properties:
        code:
          type: string
          maxLength: 50
          example: ACT-001
        title:
          type: string
          maxLength: 200
          example: Inspección del frente norte
        description:
          type: [string, 'null']
        area:
          type: string
          example: Operaciones
        location:
          type: string
          example: Frente norte
        responsible_id:
          type: string
          format: uuid
        priority:
          $ref: '#/components/schemas/ActivityPriority'
        scheduled_date:
          type: string
          format: date
        due_date:
          type: [string, 'null']
          format: date
    UpdateActivityRequest:
      type: object
      required: [title, area, location, responsible_id, priority, scheduled_date]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: [string, 'null']
        area:
          type: string
        location:
          type: string
        responsible_id:
          type: string
          format: uuid
        priority:
          $ref: '#/components/schemas/ActivityPriority'
        scheduled_date:
          type: string
          format: date
        due_date:
          type: [string, 'null']
          format: date
    ActivityResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/Activity'
    PaginatedActivitiesResponse:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      required: [current_page, per_page, total, last_page]
      properties:
        current_page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        last_page:
          type: integer
    ErrorResponse:
      type: object
      required: [success, message, errors]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
        errors:
          type: object
          additionalProperties: true
    HealthResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          required: [service, status]
          properties:
            service:
              type: string
              example: planning-service
            status:
              type: string
              example: ok